add hard coded move effect editor

+++ python/parallel.py
@@ -10,6 +10,7 @@ from learnset_reader import output_learnsets_json
 from move_reader import output_moves_json
 from arm9_reader import output_tms_json
 from grotto_odds_reader import output_grotto_odds_json
+from move_effects_table_reader import output_move_effects_table_json
 from encounter_reader import output_encounters_json
 from trdata_reader import output_trdata_json
 from item_reader import output_items_json
@@ -23,12 +24,12 @@ from matrix_reader import output_matrix_json
 
 def output(narc):	
 	narc_data = 0
-	if narc != "grotto_odds":
+	if narc != "grotto_odds" and narc != "move_effects_table":
 		file_name = f'{rom_name}/narcs/{narc}-{narc_info[narc]}.narc'
 		narc_data = ndspy.narc.NARC.fromFile(file_name)
 		os.remove(file_name)
 	else:		
-		narc_data = open(f'{rom_name}/grotto_odds.bin','rb')
+		narc_data = open(f'{rom_name}/{narc}.bin','rb')
 		# narc_data.close()
 
 	eval(f'output_{narc}_json')(narc_data, rom_name)
@@ -48,7 +49,7 @@ with open(f'{rom_name}/session_settings.json', "r") as outfile:
 narcs_to_output = ["trdata", "personal", "learnsets", "moves", "encounters", "items", "evolutions", "overworlds", "maps", "matrix"]
 # narcs_to_output = ["overworlds"]
 if narc_info["base_rom"] == "BW2":
-	narcs_to_output += ["grottos", "marts", "grotto_odds"]
+	narcs_to_output += ["grottos", "marts", "grotto_odds", "move_effects_table"]
 
 
 rom_name = narc_info["rom_name"]
